cv random forest over max_depth, not n_trees
Recommended by Andreas Mueller, who said that increasing
n_trees should always improve accuracy and that the way
to regularize was to vary the max_depth.

Andreas suggested a range of 1, 2, .. 15 for max_depth,
but this code uses 1, 2, ..., len(predictors), so that
every preditor has a change to participate in a tall,
skinny tree.